api: Move force_keyspace_cleanup endpoint to storage_service.cc
The handler needs sharded<service::storage_service>& to call
is_vnodes_cleanup_allowed(), while the rest of tasks.cc's compaction
endpoints only need the database. Move force_keyspace_cleanup (and
its two route registrations) next to the other storage_service.cc
handlers that already capture ss the same way, so that tasks.cc's
setter can stop depending on storage_service.
